+++ examples/vue/nuxt-3-catchall/pages/[...app].vue
@@ -45,7 +45,7 @@ const BUILDER_PUBLIC_API_KEY = 'f1a790f8c3204b3b8c5c1795aeac4660'; // ggignore
 const route = useRoute();
 
 // fetch builder content data
-const { data: content } = await useAsyncData('builderData', () =>
+const { data: content } = await useAsyncData(`builderData-page-${route.path}`, () =>
   fetchOneEntry({
     model: 'page',
     apiKey: BUILDER_PUBLIC_API_KEY,