+++ packages/core/src/builder.class.test.ts
@@ -499,7 +499,7 @@ describe('flushGetContentQueue', () => {
     );
   });
 
-  test("hits query url when apiEndpoint is undefined and format is 'html'", async () => {
+  test("hits content url when format is 'html'", async () => {
     const expectedFormat = 'html';
 
     const result = await builder['flushGetContentQueue'](true, [
@@ -510,35 +510,6 @@ describe('flushGetContentQueue', () => {
         userAttributes: { respectScheduling: true },
         omit: OMIT,
         fields: 'data',
-      },
-    ]);
-
-    const observerNextMock = builder.observersByKey[MODEL]?.next as jest.Mock;
-
-    expect(observerNextMock).toBeCalledTimes(1);
-    expect(observerNextMock.mock.calls[0][0][0]).toStrictEqual({
-      ...codegenOrQueryApiResult[MODEL][0],
-      variationId: expect.any(String),
-    });
-    expect(builder['makeFetchApiCall']).toBeCalledTimes(1);
-    expect(builder['makeFetchApiCall']).toBeCalledWith(
-      `https://cdn.builder.io/api/v3/query/${API_KEY}/${MODEL}?omit=${OMIT}&apiKey=${API_KEY}&fields=data&format=${expectedFormat}&userAttributes=%7B%22respectScheduling%22%3Atrue%7D&options.${MODEL}.model=%22${MODEL}%22`,
-      { headers: { Authorization: `Bearer ${AUTH_TOKEN}` } }
-    );
-  });
-
-  test("hits content url when apiEndpoint is 'content' and format is 'html'", async () => {
-    const expectedFormat = 'html';
-
-    const result = await builder['flushGetContentQueue'](true, [
-      {
-        apiEndpoint: 'content',
-        model: MODEL,
-        format: expectedFormat,
-        key: MODEL,
-        userAttributes: { respectScheduling: true },
-        omit: OMIT,
-        fields: 'data',
         limit: 10,
       },
     ]);
@@ -565,7 +536,7 @@ describe('flushGetContentQueue', () => {
     );
   });
 
-  test("hits query url when apiEndpoint is undefined and format is 'amp'", async () => {
+  test("hits content url when format is 'amp'", async () => {
     const expectedFormat = 'amp';
 
     const result = await builder['flushGetContentQueue'](true, [
@@ -576,6 +547,7 @@ describe('flushGetContentQueue', () => {
         userAttributes: { respectScheduling: true },
         omit: OMIT,
         fields: 'data',
+        limit: 10,
       },
     ]);
 
@@ -583,22 +555,29 @@ describe('flushGetContentQueue', () => {
 
     expect(observerNextMock).toBeCalledTimes(1);
     expect(observerNextMock.mock.calls[0][0][0]).toStrictEqual({
-      ...codegenOrQueryApiResult[MODEL][0],
+      ...contentApiResult.results[0],
       variationId: expect.any(String),
     });
+    expect(observerNextMock.mock.calls[0][0][1]).toStrictEqual({
+      ...contentApiResult.results[1],
+    });
+    expect(observerNextMock.mock.calls[0][0][2]).toStrictEqual({
+      ...contentApiResult.results[2],
+      variationId: expect.any(String),
+    });
+
     expect(builder['makeFetchApiCall']).toBeCalledTimes(1);
     expect(builder['makeFetchApiCall']).toBeCalledWith(
-      `https://cdn.builder.io/api/v3/query/${API_KEY}/${MODEL}?omit=${OMIT}&apiKey=${API_KEY}&fields=data&format=${expectedFormat}&userAttributes=%7B%22respectScheduling%22%3Atrue%7D&options.${MODEL}.model=%22${MODEL}%22`,
+      `https://cdn.builder.io/api/v3/content/${MODEL}?omit=data.blocks&apiKey=${API_KEY}&fields=data&format=${expectedFormat}&userAttributes=%7B%22respectScheduling%22%3Atrue%7D&limit=10&model=%22${MODEL}%22&enrich=true`,
       { headers: { Authorization: `Bearer ${AUTH_TOKEN}` } }
     );
   });
 
-  test("hits content url when apiEndpoint is 'content' and format is 'amp'", async () => {
-    const expectedFormat = 'amp';
+  test("hits content url when format is 'email'", async () => {
+    const expectedFormat = 'email';
 
     const result = await builder['flushGetContentQueue'](true, [
       {
-        apiEndpoint: 'content',
         model: MODEL,
         format: expectedFormat,
         key: MODEL,
@@ -627,138 +606,6 @@ describe('flushGetContentQueue', () => {
     expect(builder['makeFetchApiCall']).toBeCalledTimes(1);
     expect(builder['makeFetchApiCall']).toBeCalledWith(
       `https://cdn.builder.io/api/v3/content/${MODEL}?omit=data.blocks&apiKey=${API_KEY}&fields=data&format=${expectedFormat}&userAttributes=%7B%22respectScheduling%22%3Atrue%7D&limit=10&model=%22${MODEL}%22&enrich=true`,
-      { headers: { Authorization: `Bearer ${AUTH_TOKEN}` } }
-    );
-  });
-
-  test("hits query url when apiEndpoint is undefined and format is 'email'", async () => {
-    const expectedFormat = 'email';
-
-    const result = await builder['flushGetContentQueue'](true, [
-      {
-        model: MODEL,
-        format: expectedFormat,
-        key: MODEL,
-        userAttributes: { respectScheduling: true },
-        omit: OMIT,
-        fields: 'data',
-      },
-    ]);
-
-    const observerNextMock = builder.observersByKey[MODEL]?.next as jest.Mock;
-
-    expect(observerNextMock).toBeCalledTimes(1);
-    expect(observerNextMock.mock.calls[0][0][0]).toStrictEqual({
-      ...codegenOrQueryApiResult[MODEL][0],
-      variationId: expect.any(String),
-    });
-    expect(builder['makeFetchApiCall']).toBeCalledTimes(1);
-    expect(builder['makeFetchApiCall']).toBeCalledWith(
-      `https://cdn.builder.io/api/v3/query/${API_KEY}/${MODEL}?omit=${OMIT}&apiKey=${API_KEY}&fields=data&format=${expectedFormat}&userAttributes=%7B%22respectScheduling%22%3Atrue%7D&options.${MODEL}.model=%22${MODEL}%22`,
-      { headers: { Authorization: `Bearer ${AUTH_TOKEN}` } }
-    );
-  });
-
-  test("hits query url when apiEndpoint is undefined and format is 'email' and url is passed instead of userAttributes", async () => {
-    const expectedFormat = 'email';
-
-    const result = await builder['flushGetContentQueue'](true, [
-      {
-        model: MODEL,
-        format: expectedFormat,
-        key: MODEL,
-        url: '/test-page',
-        omit: OMIT,
-        fields: 'data',
-      },
-    ]);
-
-    const observerNextMock = builder.observersByKey[MODEL]?.next as jest.Mock;
-
-    expect(observerNextMock).toBeCalledTimes(1);
-    expect(observerNextMock.mock.calls[0][0][0]).toStrictEqual({
-      ...codegenOrQueryApiResult[MODEL][0],
-      variationId: expect.any(String),
-    });
-    expect(builder['makeFetchApiCall']).toBeCalledTimes(1);
-    expect(builder['makeFetchApiCall']).toBeCalledWith(
-      `https://cdn.builder.io/api/v3/query/${API_KEY}/${MODEL}?omit=${OMIT}&apiKey=${API_KEY}&fields=data&format=${expectedFormat}&userAttributes=%7B%22urlPath%22%3A%22%2Ftest-page%22%2C%22host%22%3A%22localhost%22%2C%22device%22%3A%22desktop%22%7D&options.${MODEL}.model=%22${MODEL}%22`,
-      { headers: { Authorization: `Bearer ${AUTH_TOKEN}` } }
-    );
-  });
-
-  test("hits content url when apiEndpoint is 'content' and format is 'email'", async () => {
-    const expectedFormat = 'email';
-
-    const result = await builder['flushGetContentQueue'](true, [
-      {
-        apiEndpoint: 'content',
-        model: MODEL,
-        format: expectedFormat,
-        key: MODEL,
-        userAttributes: { respectScheduling: true },
-        omit: OMIT,
-        fields: 'data',
-        limit: 10,
-      },
-    ]);
-
-    const observerNextMock = builder.observersByKey[MODEL]?.next as jest.Mock;
-
-    expect(observerNextMock).toBeCalledTimes(1);
-    expect(observerNextMock.mock.calls[0][0][0]).toStrictEqual({
-      ...contentApiResult.results[0],
-      variationId: expect.any(String),
-    });
-    expect(observerNextMock.mock.calls[0][0][1]).toStrictEqual({
-      ...contentApiResult.results[1],
-    });
-    expect(observerNextMock.mock.calls[0][0][2]).toStrictEqual({
-      ...contentApiResult.results[2],
-      variationId: expect.any(String),
-    });
-
-    expect(builder['makeFetchApiCall']).toBeCalledTimes(1);
-    expect(builder['makeFetchApiCall']).toBeCalledWith(
-      `https://cdn.builder.io/api/v3/content/${MODEL}?omit=data.blocks&apiKey=${API_KEY}&fields=data&format=${expectedFormat}&userAttributes=%7B%22respectScheduling%22%3Atrue%7D&limit=10&model=%22${MODEL}%22&enrich=true`,
-      { headers: { Authorization: `Bearer ${AUTH_TOKEN}` } }
-    );
-  });
-
-  test("hits content url when apiEndpoint is 'content' and format is 'email' and url is passed instead of userAttributes", async () => {
-    const expectedFormat = 'email';
-
-    const result = await builder['flushGetContentQueue'](true, [
-      {
-        apiEndpoint: 'content',
-        model: MODEL,
-        format: expectedFormat,
-        key: MODEL,
-        url: '/test-page',
-        omit: OMIT,
-        fields: 'data',
-        limit: 10,
-      },
-    ]);
-
-    const observerNextMock = builder.observersByKey[MODEL]?.next as jest.Mock;
-
-    expect(observerNextMock).toBeCalledTimes(1);
-    expect(observerNextMock.mock.calls[0][0][0]).toStrictEqual({
-      ...contentApiResult.results[0],
-      variationId: expect.any(String),
-    });
-    expect(observerNextMock.mock.calls[0][0][1]).toStrictEqual({
-      ...contentApiResult.results[1],
-    });
-    expect(observerNextMock.mock.calls[0][0][2]).toStrictEqual({
-      ...contentApiResult.results[2],
-      variationId: expect.any(String),
-    });
-
-    expect(builder['makeFetchApiCall']).toBeCalledTimes(1);
-    expect(builder['makeFetchApiCall']).toBeCalledWith(
-      `https://cdn.builder.io/api/v3/content/${MODEL}?omit=data.blocks&apiKey=${API_KEY}&fields=data&format=${expectedFormat}&userAttributes=%7B%22urlPath%22%3A%22%2Ftest-page%22%2C%22host%22%3A%22localhost%22%2C%22device%22%3A%22desktop%22%7D&limit=10&model=%22${MODEL}%22&enrich=true`,
       { headers: { Authorization: `Bearer ${AUTH_TOKEN}` } }
     );
   });