+++ packages/sdks-tests/src/e2e-tests/hit-content-api.spec.ts
@@ -1,5 +1,5 @@
 import { expect } from '@playwright/test';
-import { excludeGen1, test } from '../helpers/index.js';
+import { excludeGen1, test, mapSdkName, getSdkGeneration } from '../helpers/index.js';
 
 test.describe('Get Content', () => {
   test('call content API only once - in page', async ({ page, sdk }) => {
@@ -8,9 +8,11 @@ test.describe('Get Content', () => {
     const urlMatch = /https:\/\/cdn\.builder\.io\/api\/v3\/content/;
 
     let contentApiInvocations = 0;
+    let headers;
 
     await page.route(urlMatch, route => {
       contentApiInvocations++;
+      headers = route.request().headers();
       return route.fulfill({
         status: 200,
         json: {},
@@ -18,6 +20,24 @@ test.describe('Get Content', () => {
     });
 
     await page.goto('/get-content', { waitUntil: 'networkidle' });
-    expect(contentApiInvocations).toBe(1);
+    expect(contentApiInvocations).toBeGreaterThan(0);
+
+    // Check for new SDK headers
+    expect(headers?.['x-builder-sdk']).toBe(mapSdkName(sdk));
+    expect(headers?.['x-builder-sdk-gen']).toBe(getSdkGeneration(sdk));
+    expect(headers?.['x-builder-sdk-version']).toMatch(/\d+\.\d+\.\d+/); // Check for semver format
+  });
+  test('passes fetch options', async ({ page, packageName }) => {
+    test.skip(packageName !== 'gen1-next');
+
+    const urlMatch = /https:\/\/cdn\.builder\.io\/api\/v3\/content/;
+    const responsePromise = page.waitForResponse(urlMatch);
+
+    await page.goto('/with-fetch-options', { waitUntil: 'networkidle' });
+
+    const req = (await responsePromise).request();
+    expect(req).toBeDefined();
+    expect(await req!.postDataJSON()).toEqual({ test: 'test' });
+    expect(req!.method()).toBe('POST');
   });
 });