+++ packages/sdks-tests/src/specs/homepage.ts
@@ -1,6 +1,4 @@
-import { NEW_TEXT } from './helpers.js';
-
-export const CONTENT = {
+export const HOMEPAGE = {
   createdBy: 'OcOewqA7uqVVlVfqY453F8vgcc33',
   createdDate: 1644861373771,
   data: {
@@ -1078,6 +1076,3 @@ export const CONTENT = {
   firstPublished: 1644861541599,
   rev: 'ooerksctkc',
 };
-
-export const MODIFIED_HOMEPAGE = JSON.parse(JSON.stringify(CONTENT));
-MODIFIED_HOMEPAGE.data.blocks[0].children![0].component!.options.text = NEW_TEXT;