+++ packages/sdks/src/components/blocks/blocks.lite.tsx
@@ -31,13 +31,13 @@ export default function Blocks(props: BlocksProps) {
         rsc: props.context?.value?.BlocksWrapper,
         default:
           props.context?.value?.BlocksWrapper ||
-          builderContext.value.BlocksWrapper,
+          builderContext.value?.BlocksWrapper,
       })}
       BlocksWrapperProps={useTarget({
         rsc: props.context?.value?.BlocksWrapperProps,
         default:
           props.context?.value?.BlocksWrapperProps ||
-          builderContext.value.BlocksWrapperProps,
+          builderContext.value?.BlocksWrapperProps,
       })}
       classNameProp={props.className}
     >
@@ -55,7 +55,7 @@ export default function Blocks(props: BlocksProps) {
                 rsc: props.registeredComponents,
                 default:
                   props.registeredComponents ||
-                  componentsContext.registeredComponents,
+                  componentsContext?.registeredComponents,
               })}
               linkComponent={props.linkComponent}
             />